Changed e to a static variable in the Scheduler.java so that we don't have to pass e to every method.

+++ src/Schedule.java
@@ -169,7 +169,7 @@ public class Schedule {
         // Reading some general information
         m.setFullName(br.readLine());
         m.setAddress(br.readLine());
-        m.setSalary(Integer.parseInt(br.readLine()));
+        m.setSalary(Double.parseDouble(br.readLine()));
 
         // Reading the availability information
         for (int i = 0; i < 7; i++) {
@@ -189,7 +189,7 @@ public class Schedule {
         // Reading some general information
         w.setFullName(br.readLine());
         w.setAddress(br.readLine());
-        w.setWage(Integer.parseInt(br.readLine()));
+        w.setWage(Double.parseDouble(br.readLine()));
 
         // Reading the availability information
         for (int i = 0; i < 7; i++) {
